+++ CHANGELOG.md
@@ -12,6 +12,7 @@ For a full diff see [`0.17.0...main`][0.17.0...main].
 ### Added
 
 - Add option to sort books by note creation date
+- Add reading streak chart to statistics
 
 ### Changed